requirements.txt
Django==1.7.4
gunicorn==19.3.0
djangorestframework==3.1.1
django-bower==5.0.3
Pillow==2.8.1
Wand==0.4.0
django-imagekit==3.2.6
python3-memcached==1.51
drf-extensions==0.2.7
git+https://github.com/Fondation-Servir/django-mathfield.git#egg=django-mathfield
